Improve accuracy of getSizeChangeInKb
We no longer need to return -1 to signify failure, so we can improve the
precision of the variable

+++ AnkiDroid/src/main/java/com/ichi2/libanki/Collection.java
@@ -2188,8 +2188,8 @@ public class Collection {
             this.mNewSize = size;
         }
 
-        public long getSizeChangeInKb() {
-            return (mOldSize - mNewSize) / 1024;
+        public double getSizeChangeInKb() {
+            return (mOldSize - mNewSize) / 1024.0;
         }
